fix: ensure numeric conversion for stats data and improve byte formatting logic

+++ backend/src/services/StatsService.service.ts
@@ -22,7 +22,7 @@ export class StatsService {
     const [fileCount, folderCount, sizeResult, recent] = await Promise.all([
       db.select({ count: sql<number>`count(*)::int` }).from(files).where(eq(files.userId, userId)),
       db.select({ count: sql<number>`count(*)::int` }).from(folders).where(eq(folders.userId, userId)),
-      db.select({ total: sql<number>`coalesce(sum(${files.size}), 0)::bigint` }).from(files).where(eq(files.userId, userId)),
+      db.select({ total: sql<number>`coalesce(sum(${files.size}), 0)` }).from(files).where(eq(files.userId, userId)),
       db.select().from(files).where(eq(files.userId, userId)).orderBy(desc(files.createdAt)).limit(5),
     ]);
